iOS notifications are working
